+++ jkorm/src/main/kotlin/com/tang/jkorm/config/DefaultConfig.kt
@@ -1,0 +1,14 @@
+package com.tang.jkorm.config
+
+/**
+ * Default configuration
+ *
+ * @author Tang
+ */
+object DefaultConfig {
+
+    const val PAGE_NUMBER = 1L
+
+    const val PAGE_SIZE = 10L
+
+}