+++ jkorm/src/main/kotlin/com/tang/jkorm/constants/BaseMethodName.kt
@@ -61,9 +61,20 @@ object BaseMethodName {
         return method.name == COUNT
     }
 
+    private const val PAGINATE = "paginate"
+
+    fun isPaginate(method: Method): Boolean {
+        return method.name == PAGINATE
+    }
+
     fun isBaseMethod(methodName: String): Boolean {
         return when (methodName) {
-            INSERT, INSERT_SELECTIVE, UPDATE, UPDATE_SELECTIVE, DELETE, DELETE_BY_ID, SELECT, SELECT_BY_ID, COUNT  -> true
+            INSERT, INSERT_SELECTIVE,
+            UPDATE, UPDATE_SELECTIVE,
+            DELETE, DELETE_BY_ID,
+            SELECT, SELECT_BY_ID,
+            COUNT,
+            PAGINATE -> true
             else -> false
         }
     }