+++ jkorm/src/main/kotlin/com/tang/jkorm/sql/SqlProviderFactory.kt
@@ -1,0 +1,14 @@
+package com.tang.jkorm.sql
+
+import com.tang.jkorm.sql.provider.SqlProvider
+
+/**
+ * SQL provider factory
+ *
+ * @author Tang
+ */
+interface SqlProviderFactory {
+
+    fun newSqlProvider(driverClass: String): SqlProvider
+
+}