+++ jkorm/src/main/kotlin/com/tang/jkorm/sql/provider/AbstractSqlProvider.kt
@@ -1,17 +1,17 @@
-package com.tang.jkorm.sql
+package com.tang.jkorm.sql.provider
 
 import com.tang.jkorm.utils.Reflects
 
 /**
  * @author Tang
  */
-object SqlProvider {
+abstract class AbstractSqlProvider : SqlProvider {
 
-    private fun getSql(sql: StringBuilder): String {
+    override fun getSql(sql: StringBuilder): String {
         return sql.toString().lowercase()
     }
 
-    fun insert(entity: Any): String {
+    override fun insert(entity: Any): String {
         val clazz = entity::class.java
         val sql = StringBuilder()
         val idField = Reflects.getIdField(clazz)
@@ -39,7 +39,7 @@ object SqlProvider {
         return getSql(sql)
     }
 
-    private fun selectiveStrategy(any: Any?): Boolean {
+    fun selectiveStrategy(any: Any?): Boolean {
         if (any == null) {
             return false
         }
@@ -62,7 +62,7 @@ object SqlProvider {
         return map
     }
 
-    fun insertSelective(entity: Any): String {
+    override fun insertSelective(entity: Any): String {
         val clazz = entity::class.java
         val sql = StringBuilder()
         sql.append("insert into ${Reflects.getTableName(clazz)} (")
@@ -96,7 +96,7 @@ object SqlProvider {
         }
     }
 
-    fun update(entity: Any): String {
+    override fun update(entity: Any): String {
         val clazz = entity::class.java
         val sql = StringBuilder()
         val idField = Reflects.getIdField(clazz)
@@ -106,14 +106,14 @@ object SqlProvider {
             .forEach {
                 Reflects.makeAccessible(it, entity)
                 appendValue(sql, it.name, it.get(entity))
-        }
+            }
         sql.deleteCharAt(sql.length - 1)
         Reflects.makeAccessible(idField, entity)
         sql.append(" where ${idField.name} = ${idField.get(entity)}")
         return getSql(sql)
     }
 
-    fun updateSelective(entity: Any): String {
+    override fun updateSelective(entity: Any): String {
         val clazz = entity::class.java
         val sql = StringBuilder()
         sql.append("update ${Reflects.getTableName(clazz)} set ")
@@ -132,7 +132,7 @@ object SqlProvider {
         return getSql(sql)
     }
 
-    fun <T> delete(clazz: Class<T>, entity: Any): String {
+    override fun <T> delete(clazz: Class<T>, entity: Any): String {
         val sql = StringBuilder()
         val idField = Reflects.getIdField(clazz)
         Reflects.makeAccessible(idField, entity)
@@ -140,7 +140,7 @@ object SqlProvider {
         return getSql(sql)
     }
 
-    fun <T> select(clazz: Class<T>, entity: Any?): String {
+    override fun <T> select(clazz: Class<T>, entity: Any?): String {
         val sql = StringBuilder()
         sql.append("select * from ${Reflects.getTableName(clazz)}")
         if (entity == null) {
@@ -163,7 +163,7 @@ object SqlProvider {
         return getSql(sql)
     }
 
-    fun <T> count(clazz: Class<T>, entity: Any?): String {
+    override fun <T> count(clazz: Class<T>, entity: Any?): String {
         val sql = StringBuilder()
         sql.append("select count(*) from ${Reflects.getTableName(clazz)}")
         if (entity == null) {
@@ -183,6 +183,36 @@ object SqlProvider {
             }
         }
         sql.delete(sql.length - 5, sql.length)
+        return getSql(sql)
+    }
+
+    override fun <T> paginate(type: Class<T>, entity: Any?, orderBys: Array<Pair<String, Boolean>>, pageNumber: Long, pageSize: Long): String {
+        val sql = StringBuilder()
+        sql.append("select * from ${Reflects.getTableName(type)}")
+        if (entity != null) {
+            sql.append(" where ")
+            type.declaredFields.forEach {
+                Reflects.makeAccessible(it, entity)
+                val value = it.get(entity)
+                if (selectiveStrategy(value)) {
+                    sql.append(it.name).append("=")
+                    if (it.type == String::class.java) {
+                        sql.append("'").append(value).append("' and ")
+                    } else {
+                        sql.append(value).append(" and ")
+                    }
+                }
+            }
+            sql.delete(sql.length - 5, sql.length)
+            if (orderBys.isNotEmpty()) {
+                sql.append(" order by ")
+                orderBys.forEach {
+                    sql.append(it.first).append(if (it.second) " asc," else " desc,")
+                }
+                sql.deleteCharAt(sql.length - 1)
+            }
+        }
+        sql.append(" limit ").append((pageNumber - 1) * pageSize).append(",").append(pageSize)
         return getSql(sql)
     }