+++ jkorm/src/main/kotlin/com/tang/jkorm/sql/provider/mysql/MysqlSqlProvider.kt
@@ -1,0 +1,11 @@
+package com.tang.jkorm.sql.provider.mysql
+
+import com.tang.jkorm.sql.provider.AbstractSqlProvider
+
+/**
+ * MySQL SQL provider
+ *
+ * @author Tang
+ */
+class MysqlSqlProvider : AbstractSqlProvider() {
+}