+++ CHANGELOG.md
@@ -10,3 +10,4 @@
  * Support select condition query
  * Support count query
  * Support count condition query
+ * Support pagination query